Massive rewrite to thumbnail and preview generators, using higher level functionality of CHMKit.framework.
Since most of the "chmox" classes duplicated functionality found in the recently-created CHMKit.framework of the main iChm Xcode project, I changed the generator to use those classes instead.

Created a `CHMQuickLookHTMLDocument` model object that takes a `CHMDocumentFile` and `CHMLinkItem` and handles adapting the HTML data to use the `cid:` namespace. So far, it seems to be a slight improvement over the previous implementation, as I also adapt links to .css files as well as images.

I still hope to adapt the CSS usage of `url()` for styles.