Improved handling of character encodings.
The vast majority of all CHM files seem to contain HTML pages that use ISOLatin1 encoding (iso_8859_1), so when CHMQuickLookHTMLDocument creates the NSXMLDocument, if the xml document doesn't have a character encoding set, we set a default value of ISOLatin1. WinLatin1 is now the last choice, preferring UTF-8 over it if possible.